Rename Rack::Escrow to SecureEscrow
Rails Engine loader determines asset root by looking at callstack information at inheritance time. It skips over lib/rack frames when trying to determine where the gem lives causing it to fall through to too-deep frames. This was making Rails look in the active support gem directory for this gem's assets.

Renaming the gem and restructuring the library to not use the Rack:: prefix seems to be the simplest solution.